requirements.txt
fastapi
uvicorn
ibm-watson
pymongo
python-dotenv